requirements.txt
flask
flask-cors
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
